Struct state_manager::StateManager

pub struct StateManager<DB> { /* fields omitted */ }

Implementations

impl<DB> StateManager<DB> where
DB: BlockStore,

pub fn get_sectors_for_winning_post<V>(
&self,
st: &Cid,
miner_address: &Address,
rand: Randomness
) -> Result<Vec<SectorInfo>, Box<dyn StdError>> where
V: ProofVerifier,

Retrieves and generates a vector of sector info for the winning PoSt verification.

pub fn get_miner_sector_set<V>(
&self,
tipset: &Tipset,
address: &Address,
filter: &mut Option<&mut BitField>,
filter_out: bool
) -> Result<Vec<ChainSectorInfo>, Error> where
V: ProofVerifier,

pub fn miner_sector_info<V>(
&self,
address: &Address,
sector_number: &SectorNumber,
tipset: &Tipset
) -> Result<Option<SectorOnChainInfo>, Error> where
V: ProofVerifier,

pub fn precommit_info<V>(
&self,
address: &Address,
sector_number: &SectorNumber,
tipset: &Tipset
) -> Result<SectorPreCommitOnChainInfo, Error> where
V: ProofVerifier,

pub fn get_miner_info<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<MinerInfo, Box<dyn StdError>> where
V: ProofVerifier,

pub fn get_miner_deadlines<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<Deadlines, Error> where
V: ProofVerifier,

pub fn get_miner_faults<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<BitField, Box<dyn StdError>> where
V: ProofVerifier,

pub fn get_miner_recoveries<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<BitField, Box<dyn StdError>> where
V: ProofVerifier,

pub fn list_miner_actors<'a, V>(
&self,
tipset: &'a Tipset
) -> Result<Vec<Address>, Error> where
V: ProofVerifier,

impl<DB> StateManager<DB> where
DB: BlockStore,

pub fn new(bs: Arc<DB>) -> Self

pub fn new_with_subscribers(
bs: Arc<DB>,
chain_subs: Subscriber<HeadChange>
) -> Self

pub fn load_actor_state<D>(
&self,
addr: &Address,
state_cid: &Cid
) -> Result<D, Error> where
D: DeserializeOwned,

Loads actor state from IPLD Store

pub fn get_actor(
&self,
addr: &Address,
state_cid: &Cid
) -> Result<Option<ActorState>, Error>

pub fn blockstore_cloned(&self) -> Arc<DB>

pub fn blockstore(&self) -> &DB

pub fn get_network_name(&self, st: &Cid) -> Result<String, Error>

Returns the network name from the init actor state

pub fn is_miner_slashed(
&self,
addr: &Address,
state_cid: &Cid
) -> Result<bool, Error>

Returns true if miner has been slashed or is considered invalid

pub fn get_miner_work_addr(
&self,
state_cid: &Cid,
addr: &Address
) -> Result<Address, Error>

Returns raw work address of a miner

pub fn get_power(
&self,
state_cid: &Cid,
addr: &Address
) -> Result<(Claim, Claim), Error>

Returns specified actor's claimed power and total network power as a tuple

pub fn get_subscriber(&self) -> Option<Subscriber<HeadChange>>

pub fn apply_blocks<R, V, CB>(
&self,
parent_epoch: ChainEpoch,
p_state: &Cid,
messages: &[BlockMessages],
epoch: ChainEpoch,
rand: &R,
base_fee: BigInt,
callback: Option<CB>
) -> Result<(Cid, Cid), Box<dyn StdError>> where
R: Rand,
V: ProofVerifier,
CB: FnMut(Cid, &ChainMessage, ApplyRet) -> Result<(), String>,

Performs the state transition for the tipset and applies all unique messages in all blocks. This function returns the state root and receipt root of the transition.

pub async fn tipset_state<V, '_, '_>(
&'_ self,
tipset: &'_ Tipset
) -> Result<(Cid, Cid), Box<dyn StdError>> where
V: ProofVerifier,

pub fn call<V>(
&self,
message: &mut UnsignedMessage,
tipset: Option<Tipset>
) -> StateCallResult where
V: ProofVerifier,

runs the given message and returns its result without any persisted changes.

pub async fn call_with_gas<V, '_, '_, '_>(
&'_ self,
message: &'_ mut ChainMessage,
prior_messages: &'_ [ChainMessage],
tipset: Option<Tipset>
) -> StateCallResult where
V: ProofVerifier,

pub fn replay<V>(
&self,
ts: &Tipset,
mcid: &Cid
) -> Result<(UnsignedMessage, Option<ApplyRet>), Error> where
V: ProofVerifier,

returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

pub fn compute_tipset_state<V, CB>(
&self,
block_headers: &[BlockHeader],
callback: Option<CB>
) -> Result<(Cid, Cid), Box<dyn StdError>> where
V: ProofVerifier,
CB: FnMut(Cid, &ChainMessage, ApplyRet) -> Result<(), String>,

pub fn get_receipt(
&self,
tipset: &Tipset,
msg: &Cid
) -> Result<MessageReceipt, Error>

returns a message receipt from a given tipset and message cid

pub async fn wait_for_message<'a, '_>(
block_store: Arc<DB>,
subscriber: Option<Subscriber<HeadChange>>,
cid: &'_ Cid,
confidence: i64
) -> Result<(Option<Arc<Tipset>>, Option<MessageReceipt>), Error> where
DB: BlockStore + Send + Sync + 'static,

WaitForMessage blocks until a message appears on chain. It looks backwards in the chain to see if this has already happened. It guarantees that the message has been on chain for at least confidence epochs without being reverted before returning.

pub fn get_bls_public_key(
db: &Arc<DB>,
addr: &Address,
state_cid: &Cid
) -> Result<[u8; 48], Error>

Returns a bls public key from provided address

pub fn get_heaviest_balance(&self, addr: &Address) -> Result<BigInt, Error>

Return the heaviest tipset's balance from self.db for a given address

pub fn get_balance(&self, addr: &Address, cid: &Cid) -> Result<BigInt, Error>

Return the balance of a given address and state_cid

pub fn lookup_id(
&self,
addr: &Address,
ts: &Tipset
) -> Result<Option<Address>, Error>

pub fn market_balance(
&self,
addr: &Address,
ts: &Tipset
) -> Result<MarketBalance, Error>

pub async fn resolve_to_key_addr<V, '_, '_, '_>(
&'_ self,
addr: &'_ Address,
ts: &'_ Tipset
) -> Result<Address, Box<dyn StdError>> where
V: ProofVerifier,

Similar to resolve_to_key_addr in the vm crate but does not allow Actor type of addresses. Uses ts to generate the VM state.

pub fn miner_has_min_power(
&self,
addr: &Address,
ts: &Tipset
) -> Result<bool, String>

Checks power actor state for if miner meets consensus minimum requirements.