Struct state_manager::StateManager

pub struct StateManager<DB> { /* fields omitted */ }

State manager handles all interactions with the internal Filecoin actors state. This encapsulates the ChainStore functionality, which only handles chain data, to allow for interactions with the underlying state of the chain. The state manager not only allows interfacing with state, but also is used when performing state transitions.

Implementations

impl<DB> StateManager<DB> where
DB: BlockStore + Send + Sync + 'static,

pub fn get_sectors_for_winning_post<V>(
&self,
st: &Cid,
nv: NetworkVersion,
miner_address: &Address,
rand: Randomness
) -> Result<Vec<SectorInfo>, Box<dyn StdError>> where
V: ProofVerifier,

Retrieves and generates a vector of sector info for the winning PoSt verification.

pub fn get_miner_sector_set<V>(
&self,
tipset: &Tipset,
address: &Address,
filter: Option<&BitField>
) -> Result<Vec<SectorOnChainInfo>, Error> where
V: ProofVerifier,

Loads sectors for miner at given Address.

pub fn miner_sector_info<V>(
&self,
address: &Address,
sector_number: SectorNumber,
tipset: &Tipset
) -> Result<Option<SectorOnChainInfo>, Error> where
V: ProofVerifier,

Returns miner’s sector info for a given index.

pub fn precommit_info<V>(
&self,
address: &Address,
sector_number: &SectorNumber,
tipset: &Tipset
) -> Result<SectorPreCommitOnChainInfo, Error> where
V: ProofVerifier,

Returns the precommitted sector info for a miner’s sector.

pub fn get_miner_info<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<MinerInfo, Box<dyn StdError>> where
V: ProofVerifier,

Returns miner info at the given Tipset’s state.

pub fn get_miner_faults<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<BitField, Box<dyn StdError>> where
V: ProofVerifier,

Returns a bitfield of all miner’s faulty sectors.

pub fn get_miner_recoveries<V>(
&self,
tipset: &Tipset,
address: &Address
) -> Result<BitField, Box<dyn StdError>> where
V: ProofVerifier,

Returns bitfield of miner’s recovering sectors.

pub fn list_miner_actors<V>(
&self,
tipset: &Tipset
) -> Result<Vec<Address>, Error> where
V: ProofVerifier,

Lists all miners that exist in the power actor state at given Tipset.

pub fn get_miner_worker_raw(
&self,
state: &Cid,
miner_addr: &Address
) -> Result<Address, Error>

Gets miner’s worker address from state.

impl<DB> StateManager<DB> where
DB: BlockStore + Send + Sync + 'static,

pub fn new(cs: Arc<ChainStore<DB>>) -> Self

pub fn new_with_publisher(
cs: Arc<ChainStore<DB>>,
chain_subs: Publisher<HeadChange>
) -> Self

Creates a constructor that passes in a HeadChange publisher.

pub fn get_network_version(&self, epoch: ChainEpoch) -> NetworkVersion

Returns network version for the given epoch.

pub fn get_actor(
&self,
addr: &Address,
state_cid: &Cid
) -> Result<Option<ActorState>, Error>

Gets actor from given Cid, if it exists.

pub fn blockstore_cloned(&self) -> Arc<DB>

Returns the cloned Arc of the state manager’s BlockStore.

pub fn blockstore(&self) -> &DB

Returns a reference to the state manager’s BlockStore.

pub fn chain_store(&self) -> &Arc<ChainStore<DB>>

Returns reference to the state manager’s ChainStore.

pub fn get_network_name(&self, st: &Cid) -> Result<String, Error>

Returns the network name from the init actor state.

pub fn is_miner_slashed(
&self,
addr: &Address,
state_cid: &Cid
) -> Result<bool, Error>

Returns true if miner has been slashed or is considered invalid.

pub fn get_miner_work_addr(
&self,
state_cid: &Cid,
addr: &Address
) -> Result<Address, Error>

Returns raw work address of a miner given the state root.

pub fn get_power(
&self,
state_cid: &Cid,
addr: Option<&Address>
) -> Result<Option<(Claim, Claim)>, Error>

Returns specified actor’s claimed power and total network power as a tuple.

pub fn get_subscriber(&self) -> Option<Subscriber<HeadChange>>

Subscribes to the HeadChanges observed by the state manager.

pub fn apply_blocks<R, V, CB>(
self: &Arc<Self>,
parent_epoch: ChainEpoch,
p_state: &Cid,
messages: &[BlockMessages],
epoch: ChainEpoch,
rand: &R,
base_fee: BigInt,
callback: Option<CB>,
tipset: &Arc<Tipset>
) -> Result<(Cid, Cid), Box<dyn StdError>> where
R: Rand,
V: ProofVerifier,
CB: FnMut(&Cid, &ChainMessage, &ApplyRet) -> Result<(), String>,

Performs the state transition for the tipset and applies all unique messages in all blocks. This function returns the state root and receipt root of the transition.

pub async fn tipset_state<V>(
self: &Arc<Self>,
tipset: &Arc<Tipset>
) -> Result<(Cid, Cid), Box<dyn StdError>> where
V: ProofVerifier,

Returns the pair of (parent state root, message receipt root). This will either be cached or will be calculated and fill the cache. Tipset state for a given tipset is guaranteed not to be computed twice.

pub async fn call<V>(
self: &Arc<Self>,
message: &mut UnsignedMessage,
tipset: Option<Arc<Tipset>>
) -> Result<InvocResult, Error> where
V: ProofVerifier,

runs the given message and returns its result without any persisted changes.

pub async fn call_with_gas<V>(
self: &Arc<Self>,
message: &mut ChainMessage,
prior_messages: &[ChainMessage],
tipset: Option<Arc<Tipset>>
) -> Result<InvocResult, Error> where
V: ProofVerifier,

Computes message on the given Tipset state, after applying other messages and returns the values computed in the VM.

pub async fn replay<V>(
self: &Arc<Self>,
ts: &Arc<Tipset>,
mcid: Cid
) -> Result<(UnsignedMessage, ApplyRet), Error> where
V: ProofVerifier,

Replays the given message and returns the result of executing the indicated message, assuming it was executed in the indicated tipset.

pub async fn get_lookback_tipset_for_round<V>(
self: &Arc<Self>,
tipset: Arc<Tipset>,
round: ChainEpoch
) -> Result<(Arc<Tipset>, Cid), Error> where
V: ProofVerifier,

Gets lookback tipset for block validations.

pub fn eligible_to_mine(
self: &Arc<Self>,
address: &Address,
base_tipset: &Tipset,
lookback_tipset: &Tipset
) -> Result<bool, Error>

Checks the eligibility of the miner. This is used in the validation that a block’s miner has the requirements to mine a block.

pub async fn miner_get_base_info<V: ProofVerifier, B: Beacon>(
self: &Arc<Self>,
beacon: &BeaconSchedule<B>,
key: &TipsetKeys,
round: ChainEpoch,
address: Address
) -> Result<Option<MiningBaseInfo>, Box<dyn StdError>>

Get’s a miner’s base info from state, based on the address provided.

pub async fn compute_tipset_state<V, CB: 'static>(
self: &Arc<Self>,
tipset: &Arc<Tipset>,
callback: Option<CB>
) -> Result<(Cid, Cid), Error> where
V: ProofVerifier,
CB: FnMut(&Cid, &ChainMessage, &ApplyRet) -> Result<(), String> + Send,

Performs a state transition, and returns the state and receipt root of the transition.

pub async fn get_receipt(
&self,
tipset: &Tipset,
msg: &Cid
) -> Result<MessageReceipt, Error>

Returns a message receipt from a given tipset and message cid.

pub async fn wait_for_message(
self: &Arc<Self>,
msg_cid: Cid,
confidence: i64
) -> Result<(Option<Arc<Tipset>>, Option<MessageReceipt>), Error> where
DB: BlockStore + Send + Sync + 'static,

WaitForMessage blocks until a message appears on chain. It looks backwards in the chain to see if this has already happened. It guarantees that the message has been on chain for at least confidence epochs without being reverted before returning.

pub fn get_bls_public_key(
db: &DB,
addr: &Address,
state_cid: &Cid
) -> Result<[u8; 48], Error>

Returns a bls public key from provided address

pub async fn get_heaviest_balance(
&self,
addr: &Address
) -> Result<BigInt, Error>

Return the heaviest tipset’s balance from self.db for a given address

pub fn get_balance(&self, addr: &Address, cid: &Cid) -> Result<BigInt, Error>

Return the balance of a given address and state_cid

pub fn lookup_id(
&self,
addr: &Address,
ts: &Tipset
) -> Result<Option<Address>, Error>

Looks up ID Address from the state at the given Tipset.

pub fn market_balance(
&self,
addr: &Address,
ts: &Tipset
) -> Result<MarketBalance, Error>

Retrieves market balance in escrow and locked tables.

pub async fn resolve_to_key_addr<V>(
self: &Arc<Self>,
addr: &Address,
ts: &Arc<Tipset>
) -> Result<Address, Box<dyn StdError>> where
V: ProofVerifier,

Similar to resolve_to_key_addr in the vm crate but does not allow Actor type of addresses. Uses ts to generate the VM state.

pub fn miner_has_min_power(
&self,
addr: &Address,
ts: &Tipset
) -> Result<bool, Box<dyn StdError>>

Checks power actor state for if miner meets consensus minimum requirements.

pub async fn validate_chain<V: ProofVerifier>(
self: &Arc<Self>,
ts: Arc<Tipset>,
height: i64
) -> Result<(), Box<dyn StdError>>

pub fn get_circulating_supply(
self: &Arc<Self>,
height: ChainEpoch,
state_tree: &StateTree<'_, DB>
) -> Result<TokenAmount, Box<dyn StdError>>

Retrieves total circulating supply on the network.

pub fn get_market_state(&self, ts: &Tipset) -> Result<State, Error>

Return the state of Market Actor.